sim/requirements.ubuntu.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
beniget==0.4.2.post1
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blobconverter==1.4.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
coacd==1.0.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distinctipy==1.3.4
distro==1.9.0
docker==7.1.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.21.2
durationpy==0.10
ecdsa==0.19.1
empy==4.2
etils==1.13.0
exceptiongroup==1.3.0
executing==2.2.0
fake-useragent==2.2.0
fastapi==0.116.1
fasteners==0.19
fastjsonschema==2.21.1
ffmpy3==0.2.4
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.2.10
fonttools==4.59.0
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.7.0
fs==2.4.16
fsspec==2025.7.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
genesis-world==0.4.3
glfw==2.9.0
google-auth==2.49.1
google-auth-oauthlib==1.2.2
google-genai==1.66.0
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
imread_from_url==0.1.3
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.10.0
Keras-Preprocessing==1.1.2
keyring==25.6.0
kiwisolver==1.4.8
lark==1.2.2
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.7
lazy_loader==0.4
libclang==18.1.1
libigl==2.6.1
lightning-utilities==0.14.3
llvmlite==0.44.0
lockfile==0.12.2
lxml==6.0.0
lz4==4.4.4
macaroonbakery==1.3.4
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcap==1.3.0
mcap-ros1-support==0.7.3
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==10.7.0
moviepy==2.2.1
mpi4py==4.1.0
mpmath==1.3.0
mujoco==3.3.4
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.5
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.2
numpy==2.1.3
numpy-stl==3.2.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
olefile==0.47
ompl==1.7.0
onnx==1.18.0
onnxruntime==1.22.1
openai==1.97.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
OpenEXR==3.3.4
opt_einsum==3.4.0
optree==0.16.0
overrides==7.7.0
packaging==25.0
pafy==0.5.5
pandas==2.3.1
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.3.8
plotly==5.24.1
pluggy==1.6.0
ply==3.11
pooch==1.8.2
proglog==0.1.12
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptflops @ git+https://github.com/sovrasov/flops-counter.pytorch.git@933f36edc52020aba58dfcfe6a5c0054a779dff3
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycollada==0.9.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydocstyle==6.3.0
pydot==4.0.1
pyflakes==3.4.0
PyGEL3D==0.6.1
pyglet==2.1.6
pygltflib==1.16.0
Pygments==2.19.2
PyJWT==2.10.1
pymacaroons==0.13.0
pymeshlab==2023.12.post3
PyNaCl==1.5.0
PyOpenGL==3.1.9
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyRFC3339==1.1
pyrsistent==0.20.0
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-json-logger==3.3.0
python-utils==3.9.1
pythran==0.18.0
pytorch-lightning==2.5.2
pytube==15.0.0
PyTurboJPEG==1.8.2
pytz==2025.2
pyusb==1.3.1
pyvista==0.45.3
pyxdg==0.28
PyYAML==6.0.2
pyzed==1.3.0
pyzmq==27.0.0
Qt.py==1.4.6
referencing==0.36.2
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
roman==5.1
rosbags==0.10.10
rpds-py==0.26.0
rsa==4.9.1
rtree==1.4.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
scooby==0.10.1
SecretStorage==3.3.3
Send2Trash==1.8.3
sentry-sdk==2.33.0
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.1
stringzilla==3.12.5
sympy==1.14.0
taichi==1.7.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
terminado==0.18.1
test_tube==0.7.5
tetgen==0.8.2
texttable==1.7.0
thop==0.1.1.post2209072238
tifffile==2025.6.11
timm==1.0.17
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch  # pinned to nightly cu128 at runtime for Blackwell (RTX 50xx) - see setup.sh
torchmetrics==1.7.4
torchvision  # see above
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.7.1
triton==3.3.1
typer==0.12.5
types-pyside2==5.15.2.1.7
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ufoLib2==0.18.1
unicodedata2==16.0.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
vtk==9.4.2
wadllib==2.0.0
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xmltodict==0.14.2
yarl==1.20.1
youtube-dl==2021.12.17
z3-solver==4.15.1.0
zipp==3.23.0
zstandard==0.23.0
